+++ src/Commands/Server/Member.ts
@@ -3,6 +3,24 @@ import { Message, GuildMember, Activity } from 'discord.js';
 import Embed from '../../Structures/Embed';
 import { formatDate } from '../../lib/Utility/Date';
 
+const formatPresence = (activities: Activity[]) => {
+    const push: string[] = [];
+    for(const activity of activities) {
+        switch(activity.type) {
+            case 'CUSTOM_STATUS':
+                push.push(`${activity.emoji ?? ' '}\`\`${activity.state}\`\``); break;
+            case 'LISTENING':
+                push.push(`Listening to ${activity.details} - ${activity.state ?? 'N/A'} on ${activity.name}.`); break;
+            case 'PLAYING':
+                push.push(`Playing *${activity.name}*.`); break;
+            default:
+                console.log(activity);
+        }
+    }
+
+    return push.join('\n');
+}
+
 export default class extends Command {
     constructor() {
         super(
@@ -12,9 +30,9 @@ export default class extends Command {
             ],
             [ /* No extra perms needed */ ],
             {
-                name: 'user',
+                name: 'member',
                 folder: 'Server',
-                aliases: [ 'userinfo', 'whois' ],
+                aliases: [ 'memberinfo', 'whois' ],
                 args: [0, 1],
                 guildOnly: true
             }
@@ -26,54 +44,30 @@ export default class extends Command {
             return message.channel.send(Embed.fail('Too many people mentioned!'));
         }
 
-        if(!/(<@!)?\d{17,19}>?/.test(args[0])) {
-            return message.channel.send(Embed.fail(`
-            No guild member mentioned and no user ID provided.
-            `));
-        }
-
         let member: GuildMember;
         try {
             member = await message.guild.members.fetch(args[0].replace(/[^\d]/g, ''));
         } catch {
-            return message.channel.send(Embed.fail('Invalid ID provided or member mentioned!'));
+            member = message.member;
         }
 
         const embed = Embed.success()
             .setDescription(`
             ${member} on *${member.guild.name}*.
-            ${this.formatPresence(member.presence.activities)}
+            ${formatPresence(member.presence.activities)}
             
             Permissions: 
             \`\`${member.permissions.toArray().join(', ')}\`\`
             `)
             .setThumbnail(member.user.displayAvatarURL())
             .addField('**Username:**',   member.user.username, true)
-            .addField('**Role Color:**', member.displayHexColor, true)
             .addField('**ID:**',         member.id, true)
+            .addField('**Role Color:**', member.displayHexColor, true)
             .addField('**Discrim:**',    `#${member.user.discriminator}`, true)
             .addField('**Nickname:**',   member.nickname ?? 'None', true)
             .addField('**Bot:**',        member.user.bot ? 'Yes' : 'No', true)
             .addField('**Created:**',     formatDate('MMMM Do, YYYY hh:mm:ss A t', member.user.createdAt), false)
         
         return message.channel.send(embed);
-    }
-
-    formatPresence(activities: Activity[]) {
-        const push: string[] = [];
-        for(const activity of activities) {
-            switch(activity.type) {
-                case 'CUSTOM_STATUS':
-                    push.push(`${activity.emoji ?? ' '}\`\`${activity.state}\`\``); break;
-                case 'LISTENING':
-                    push.push(`Listening to ${activity.details} - ${activity.state ?? 'N/A'} on ${activity.name}.`); break;
-                case 'PLAYING':
-                    push.push(`Playing *${activity.name}*.`); break;
-                default:
-                    console.log(activity);
-            }
-        }
-
-        return push.join('\n');
     }
 }